LazyPlatform/checker/requirements.txt
randcrack==0.2.0
pwntools==4.8.0
pycryptodome==3.15.0